backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.0
redis==5.1.0
aiosmtplib==3.0.1
jinja2==3.1.4
python-dotenv==1.0.1
groq==0.11.0
reportlab==4.2.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
PyMuPDF==1.24.0
pdfplumber==0.11.0
numpy==1.26.4
# WhatsApp via Whapi (usa httpx, ya incluido arriba)
google-auth==2.35.0
google-auth-oauthlib==1.2.1
